requirements.txt
python-binance>=1.0.0
python-dotenv>=1.0.0
